Add files via upload
I imported the dataset to Power BI then I removed the gender encoded , retention status encoded and location encoded as I was not interested in using them. With this dashboard I wanted to see how likely is it that a customer will buy a product after viewing it , how satisfied is the customer and which gender is the target market basically or which gender purchases more. I was also able to tell that females tend to spend more than males as 51.62% of the total number of products purchased are by females, and also people residing in-between Urban and Rural areas are the ones making more purchases , now that can even have an effect on the location of the store in the near future. If it is located in Urban areas then it may wanna move open a branch or relocate where there are more buyers.